requirements.txt
python-fasthtml>=0.1.0
google-genai>=0.8.0
python-multipart>=0.0.6
pathlib2>=2.3.7
